requirements.txt
# Core dependencies
requests
pandas
numpy
pyyaml
python-dotenv

# Dashboard & Visualization
streamlit

# Optional future tools
slack_sdk
chromadb
fastapi
uvicorn
